refactor:componentsのリファクタリング粗方完了
