+++ components/molecules/ActionList.tsx
@@ -1,0 +1,64 @@
+import React from 'react';
+import { Button, Box, List, ListItem } from '@mui/material';
+
+import LogoutButton from '../atoms/LogoutButton';
+
+import { ActionListProps } from '../../types/index';
+
+export const ActionList: React.FC<ActionListProps> = ({ 
+  anchor, 
+  onAction, 
+  handleOpenModal, 
+  physicalRecoveryIntervalId, 
+  statDecreaseIntervalId 
+}) => {
+  return (
+    <Box
+      sx={{ 
+        width: anchor === 'top' || anchor === 'bottom' ? 'auto' : 250,
+        display: 'flex',
+        flexDirection: anchor === 'top' || anchor === 'bottom' ? 'row' : 'column',
+        flexWrap: 'wrap',
+        backgroundColor: "#FFFBE8"
+      }}
+      role="presentation"
+    >
+      <List sx={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between', padding: "50px 0", width: '90%', margin:"0 auto" }}>
+        {['ご飯', 'なでる', '遊ぶ'].map((text) => (
+          <ListItem key={text} disablePadding sx={{ width: 'auto' }}>
+            <Button
+              variant="contained"
+              color="secondary"
+              sx={{ color: 'white', fontSize: "24px", zIndex: 1000 }}
+              onClick={() => {
+                if (text === 'ご飯') {
+                  onAction('feed');
+                } else if (text === 'なでる') {
+                  onAction('stroke');
+                } else if (text === '遊ぶ') {
+                  onAction('play');
+                }
+              }}
+            >
+              {text}
+            </Button>
+          </ListItem>
+        ))}
+        <Button
+          variant="contained"
+          color="primary"
+          sx={{ color: 'white', fontSize: "24px", zIndex: 1000 }}
+          onClick={handleOpenModal}
+        >
+          ステータス
+        </Button>
+        <LogoutButton  
+          physicalRecoveryIntervalId={physicalRecoveryIntervalId}
+          statDecreaseIntervalId={statDecreaseIntervalId} 
+        />
+      </List>
+    </Box>
+  );
+};
+
+export default ActionList;
